app/requirements.txt
os
pandas
sqlalchemy
python-dotenv
streamlit